Implement a "stay within bounds" mode
This creates a new control mode that operates to control 0 torque
(modulo feedforward) when the position is between two bounds.  Outside
of that region, the normal PID controller is used to force the
position back to one of the bounds.

One or both (not sure why) of the bounds can be nan to create
one-sided regions.  A feedforward torque can also be used which will
apply in the "within" region.