Track control position as an int64 with more precision
Previously, control_position was a float measured in post-gearbox
terms.  This had a number of problems, the biggest was that velocity
control was problematic as you got more than a revolution or two away
from 0.  Since the internal control loop needs to advance the control
position every 1/40000s, floats did not have enough resolution to
actually cause things to advance with any precision.

Now, the control_position is measured in terms of 1/65536 of the
pre-gearbox integer position.  This allows accurate velocity control
down to at least 0.0001 rev/s pre-gearbox uniformly across the entire
range of operation.  Also, the control loop is slightly restructured
so that position is calculated relative to the control point in
integer terms, then converted to floats.  This allows operation across
the wraparound point of the pre-gearbox int32_t position.

As of now, you can have resolution as good as the encoder when
commanding positions out to +-2048.0, and "decent" resolution all the
way out to 32768 which is where the wraparound point is.

Tests are updated to show that we can now control low velocity from
any start position, and that wraparound is possible.