Implement a cooldown period when removing power
This leaves the motor in a "brake" configuration for a fixed number of
cycles before disabling the motor driver.  It is intended to reduce DC
bus transients when stopping from high current.

If the DRV8323 experiences an over-current fault, it internally uses a
different mechanism to mitigate this by overriding the drive strength
to a low value.  We can't do that on the fly easily, so do this
instead.